Allow removal of incompatible apps
Apps which are not in the user's current channel are currently not shown
in the store. This is a problem for scopes, which offer no way to
be uninstalled through the UT interface.

This commit shows all apps installed on the device on the "installed
apps" page. It also adds a "Remove" option for apps which are not
compatible. Pressing the "Remove" button confirms that you want to
remove the app, removes the app, then closes the page.